Change chracter slash command to button, refactor
Resolves "Persistent /character button in main interface" #19
- ExplorationActionView — row 3, custom_id="action:character". Excluded
  from the disable-on-turn-close loop (always enabled).
- CombatActionView — row 1, custom_id="combat:character". The existing
  disable loop only targets combat:act, so this button already stays
enabled when rounds are closed.
- cogs/character_views.py (new) — owns EquipMenuView, EquipSelectView,
  AccessorySlotView, UnequipView, _character_sheet, _find_character,
_SLOT_LABELS
- cogs/session.py — now purely slash commands (/help, /turn, /round,
  /abscond, /say, /emote, /oracle, /strife, /status); dropped /character
and all view/helper code; keeps its own local _find_character since it
still needs it for the slash commands
- cogs/action_buttons.py — lazy imports updated from cogs.session →
  cogs.character_views
- test_startup.py — removed "character" from EXPECTED_COMMANDS
- test_interactive.py — test_character_command_responds now prompts
  "Click the 'My Character' button" instead of "Run /character", with
updated fail message; the DM polling logic is identical since the
behavior is the same
