+++ bot.py
@@ -151,7 +151,7 @@ async def start_webui():
 async def main():
     async with bot:
         await bot.load_extension("cogs.arrive")
-        await bot.load_extension("cogs.session")
+        await bot.load_extension("cogs.slash_commands")
         await bot.load_extension("cogs.dm_commands")
         await bot.load_extension("cogs.timer")
         await bot.load_extension("cogs.action_buttons")
